Refactor Shaders class to different classes and add documentation
