+++ src/main/kotlin/graphics/scenery/backends/ShaderNotFoundException.kt
@@ -2,4 +2,9 @@ package graphics.scenery.backends
 
 import java.io.FileNotFoundException
 
+/**
+ * Exception to be thrown if a shader file cannot be located.
+ *
+ * @author Ulrik Guenther <hello@ulrik.is>
+ */
 class ShaderNotFoundException(description: String): FileNotFoundException(description)
