+++ src/main/kotlin/graphics/scenery/backends/SourceSPIRVPriority.kt
@@ -1,0 +1,9 @@
+package graphics.scenery.backends
+
+/**
+ * Enums to indicate whether the source code or SPIRV bytecode in a [ShaderPackage]
+ * should take precedence.
+ *
+ * @author Ulrik Guenther <hello@ulrik.is>
+ */
+enum class SourceSPIRVPriority { SourcePriority, SPIRVPriority }
